Change the Mandelbrot set image to 8 bits
From 16 bits. 16 bits PGM images are not fully supported by all
tools. 8 bits gives 256 iterations which is enough to generate a
nice looking Mandelbrot set image with high enough "iteration
resolution".

+++ src/MandelbrotSet/include/MandelbrotSet/mandelbrot_set.h
@@ -39,6 +39,6 @@ struct MBROT_Range
 void MBROT_generate_mandelbrot_set(
     struct IMG_Image *image,
     const struct MBROT_Range *range,
-    uint16_t max_iterations);
+    uint8_t max_iterations);
 
 #endif /* MANDELBROTSET_MANDELBROTSET_H */
